+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Clearer error when it can't find matching assay name.
 - Downgrade dependency for compatibility with HuBMAP commons.
 - Directory structure for scatacseq.
+- Link to the yaml for both directory and metadata schemas.
 
 ## v0.0.8 - 2021-02-10
 - Update CODEX directory structure